move benchmark library code to benches scope
this way it's not picked up by our code coverage
